requirements.txt
Flask==2.0.1
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
torch==1.9.0
transformers==4.6.1
sqlalchemy==1.4.22
pytest==6.2.4
black==21.7b0
flake8==3.9.2
mypy==0.910

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
langchain==0.0.350
langchain-community==0.0.10
openai==1.3.0
anthropic==0.7.0

# Vector Database
chromadb==0.4.18
faiss-cpu==1.7.4
pinecone-client==2.2.4

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0
httpx==0.25.2
celery==5.3.4